Refactor TrainingHub callback E2E tests and Flatten callback-tests into sdk_tests
